refactor: file structure of the routing system
