+++ src/pages/CustomerSearch/CustomerSearchPage.tsx
@@ -3,7 +3,7 @@ import { Link, useSearch } from "@tanstack/react-router";
 import { createColumnHelper } from "@tanstack/react-table";
 import { useTranslation } from "react-i18next";
 
-import { customerSearchRoute } from "../../routes";
+import { searchCustomersRoute } from "../../routes/routing/customers/searchCustomers";
 import Protector from "../../routes/Protector";
 import ModuleTable, {
   type ColumnVisibilityGraph,
@@ -26,7 +26,7 @@ const DateColumns = ["DateOfbirth", "LicenseExpiryDate"];
 function CustomerSearchPage() {
   const { t } = useTranslation();
 
-  const search = useSearch({ from: customerSearchRoute.id });
+  const search = useSearch({ from: searchCustomersRoute.id });
   const { searchFilters, pageNumber, size } =
     normalizeCustomerListSearchParams(search);
