+++ src/pages/VehiclesSearch/VehiclesSearchPage.tsx
@@ -2,7 +2,7 @@ import { useCallback, useMemo } from "react";
 import { Link, useSearch } from "@tanstack/react-router";
 import { createColumnHelper } from "@tanstack/react-table";
 
-import { vehiclesSearchRoute } from "../../routes";
+import { searchVehiclesRoute } from "../../routes/routing/vehicles/searchVehicles";
 import Protector from "../../routes/Protector";
 import ModuleTable, {
   type ColumnVisibilityGraph,
@@ -22,7 +22,7 @@ import { normalizeVehicleListSearchParams } from "../../utils/normalize-search-p
 const columnHelper = createColumnHelper<TVehicleListItemParsed>();
 
 function VehiclesSearchPage() {
-  const search = useSearch({ from: vehiclesSearchRoute.id });
+  const search = useSearch({ from: searchVehiclesRoute.id });
   const { pageNumber, size, searchFilters } =
     normalizeVehicleListSearchParams(search);
