+++ src/routes/routing/__root.tsx
@@ -1,0 +1,13 @@
+import { createRouteConfig, Outlet } from "@tanstack/react-router";
+
+import AppShellLayout from "../../components/AppShellLayout";
+
+export const rootRoute = createRouteConfig({
+  component: () => {
+    return (
+      <AppShellLayout>
+        <Outlet />
+      </AppShellLayout>
+    );
+  },
+});
