+++ src/routes/routing/agreements/index.ts
@@ -1,0 +1,3 @@
+import { rootRoute } from "../__root";
+
+export const agreementsRoute = rootRoute.createRoute({ path: "agreements" });
