+++ src/routes/routing/customers/index.ts
@@ -1,0 +1,3 @@
+import { rootRoute } from "../__root";
+
+export const customersRoute = rootRoute.createRoute({ path: "customers" });
