+++ src/routes/routing/reservations/index.ts
@@ -1,0 +1,5 @@
+import { rootRoute } from "../__root";
+
+export const reservationsRoute = rootRoute.createRoute({
+  path: "reservations",
+});
