+++ src/routes/routing/reservations/viewReservation.ts
@@ -1,0 +1,50 @@
+import { lazy } from "@tanstack/react-router";
+import { z } from "zod";
+
+import { reservationsRoute } from ".";
+import { queryClient as qc } from "../../../App";
+import { fetchRentalRateSummaryAmounts } from "../../../api/summary";
+
+import { getAuthToken } from "../../../utils/authLocal";
+import { reservationQKeys } from "../../../utils/query-key";
+
+export const viewReservationRoute = reservationsRoute.createRoute({
+  path: "$reservationId",
+  validateSearch: (search) =>
+    z.object({ tab: z.string().optional() }).parse(search),
+  preSearchFilters: [() => ({ tab: "summary" })],
+  onLoad: async ({ params: { reservationId } }) => {
+    const auth = getAuthToken();
+
+    if (auth) {
+      const promises = [];
+      // get summary
+      const summaryKey = reservationQKeys.summary(reservationId);
+      if (!qc.getQueryData(summaryKey)) {
+        promises.push(
+          qc.prefetchQuery({
+            queryKey: summaryKey,
+            queryFn: () =>
+              fetchRentalRateSummaryAmounts({
+                clientId: auth.profile.navotar_clientid,
+                userId: auth.profile.navotar_userid,
+                accessToken: auth.access_token,
+                module: "reservations",
+                referenceId: reservationId,
+              }),
+          })
+        );
+      }
+
+      await Promise.all(promises);
+    }
+    return {};
+  },
+  component: lazy(
+    () => import("../../../pages/ReservationView/ReservationViewPage")
+  ),
+  parseParams: (params) => ({
+    reservationId: z.string().parse(params.reservationId),
+  }),
+  stringifyParams: (params) => ({ reservationId: `${params.reservationId}` }),
+});
