+++ src/routes/routing/vehicles/index.ts
@@ -1,0 +1,3 @@
+import { rootRoute } from "../__root";
+
+export const vehiclesRoute = rootRoute.createRoute({ path: "vehicles" });
